Fix tokenize invocations
In xsl 1.0 or rather in exslt the tokenize function would treat the
second argument as a list of separator characters whereas in xslt 2.0
the second argument is treated as a single separator. So to properly
move this to xslt 2.0 the second agrgument has been converted to a
coresponding regexp.

This should fix many of the remaing issues with isbn numbers,
multiword emphasis, etc.