Make sure empty elements aren't counted for em handling
When handling em or grade change announcements we need to know how
many words are in an em. This is done with tokenize (on white space
and '-') and count. However if there is a trailing '-' you get a
sequence with an empty element and falsly believe that you have more
than one word. So a filter after the tokenize now cleans out the empty
elements. This fixes #1518.