Handle noterefs before punctuation
This fixes #1264

The commit also fixes some of the test cases with respect to a fix
that was done to the Braille tables where a hyphen mark is now
inserted after a dash. 

Also the noteref tests and the dash tests are moved to separate files.

And finally there is now a function to test if a context is a block
element. This replaces the local-name()=('h1', ..., 'p') construct